Fixes
Fixed: Cooking skill in the skill didn't show anything if using Yet Another Cooking Skill with The Love of Cooking integrated.
Fixed: VPP skill menu needed a reload to show skills correctly when used with Better Game Menu
Added: Option to choose if professions should be required for pillars or pillars for professions.